requirements.txt
gitpython
dotenv
jira
seaborn
matplotlib
pandas
imbalanced-learn
tqdm